Covid Data Explorer refactor. Decoupled view from data table
