+++ charts/__tests__/CovidDataExplorer.test.tsx
@@ -4,7 +4,7 @@ import { CovidDataExplorer } from "../covidDataExplorer/CovidDataExplorer"
 import { CovidQueryParams } from "charts/covidDataExplorer/CovidChartUrl"
 import { testData } from "../../test/fixtures/CovidTestData"
 import { csvParse } from "d3-dsv"
-import { parseCovidRow } from "charts/covidDataExplorer/CovidDataUtils"
+import { parseCovidRow } from "charts/covidDataExplorer/CovidExplorerTable"
 import React from "react"
 import { shallow } from "enzyme"
 import { ParsedCovidCsvRow } from "charts/covidDataExplorer/CovidTypes"
